+++ src/main/java/com/verdantartifice/thaumicwonders/common/items/armor/ItemVoidFortressArmor.java
@@ -13,6 +13,7 @@ import net.minecraft.entity.EntityLivingBase;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.init.SoundEvents;
 import net.minecraft.inventory.EntityEquipmentSlot;
+import net.minecraft.item.EnumRarity;
 import net.minecraft.item.ItemArmor;
 import net.minecraft.item.ItemStack;
 import net.minecraft.util.DamageSource;
@@ -160,5 +161,10 @@ public class ItemVoidFortressArmor extends ItemArmor implements ISpecialArmor, I
         } else {
             return false;
         }
+    }
+    
+    @Override
+    public EnumRarity getRarity(ItemStack stack) {
+        return EnumRarity.EPIC;
     }
 }